+++ src/Libraries Projects/Milochau.Core/Infrastructure/Converters/DetailedHealthReport.cs
@@ -43,9 +43,7 @@ namespace Milochau.Core.Infrastructure.Converters
 
                 if (item.Value.Exception != null)
                 {
-                    var message = item.Value.Exception?
-                        .Message
-                        .ToString();
+                    var message = item.Value.Exception?.Message;
 
                     entry.Exception = message;
                     entry.Description = item.Value.Description ?? message;