+++ src/Libraries Projects/Milochau.Core/Infrastructure/Hosting/CoreOptionsFactory.cs
@@ -18,7 +18,6 @@ namespace Milochau.Core.Infrastructure.Hosting
         private const string hostNameKey = "HOST";
         private const string keyVaultVaultKey = "KEYVAULT_VAULT";
         private const string appConfigEndpointKey = "APPCONFIG_ENDPOINT";
-        private const string appConfigConnectionStringKey = "APPCONFIG_CONNECTIONSTRING";
 
         /// <summary>Gets a new <see cref="CoreHostOptions"/> and setup fallback values</summary>
         /// <param name="configuration">Configuration</param>
@@ -38,38 +37,26 @@ namespace Milochau.Core.Infrastructure.Hosting
             configuration.Bind(CoreHostOptions.DefaultConfigurationSection, hostOptions);
 
             hostOptions.Application.ApplicationName = hostOptions.Application.ApplicationName
-                ?? configuration[$"{GenericHostConfigurationPrefix}{applicationNameKey}"]
-                ?? configuration[$"{WebHostConfigurationPrefix}{applicationNameKey}"]
-                ?? configuration[applicationNameKey];
+                ?? GetValueFromConfiguration(configuration, applicationNameKey);
             hostOptions.Application.EnvironmentName = hostOptions.Application.EnvironmentName
-                ?? configuration[$"{GenericHostConfigurationPrefix}{environmentNameKey}"]
-                ?? configuration[$"{WebHostConfigurationPrefix}{environmentNameKey}"]
-                ?? configuration[environmentNameKey]
+                ?? GetValueFromConfiguration(configuration, environmentNameKey)
                 ?? ApplicationHostEnvironment.DevelopmentEnvironmentName;
             hostOptions.Application.HostName = hostOptions.Application.HostName
-                ?? configuration[$"{GenericHostConfigurationPrefix}{hostNameKey}"]
-                ?? configuration[$"{WebHostConfigurationPrefix}{hostNameKey}"]
-                ?? configuration[hostNameKey]
+                ?? GetValueFromConfiguration(configuration, hostNameKey)
                 ?? ApplicationHostEnvironment.LocalHostName;
 
             hostOptions.KeyVault.Vault = hostOptions.KeyVault.Vault
-                ?? configuration[$"{GenericHostConfigurationPrefix}{keyVaultVaultKey}"]
-                ?? configuration[$"{WebHostConfigurationPrefix}{keyVaultVaultKey}"]
-                ?? configuration[keyVaultVaultKey];
+                ?? GetValueFromConfiguration(configuration, keyVaultVaultKey);
 
             hostOptions.AppConfig.Endpoint = hostOptions.AppConfig.Endpoint
-                ?? configuration[$"{GenericHostConfigurationPrefix}{appConfigEndpointKey}"]
-                ?? configuration[$"{WebHostConfigurationPrefix}{appConfigEndpointKey}"]
-                ?? configuration[appConfigEndpointKey];
+                ?? GetValueFromConfiguration(configuration, appConfigEndpointKey);
         }
 
         /// <summary>Gets current environment name from environment variables</summary>
         /// <remarks>This method only gets environment name from environment variables. It shall only be used in application initialization.</remarks>
         public static string GetCurrentEnvironmentFromEnvironmentVariables()
         {
-            return Environment.GetEnvironmentVariable($"{GenericHostConfigurationPrefix}{environmentNameKey}")
-                ?? Environment.GetEnvironmentVariable($"{WebHostConfigurationPrefix}{environmentNameKey}")
-                ?? Environment.GetEnvironmentVariable(environmentNameKey)
+            return GetValueFromEnvironment(environmentNameKey)
                 ?? ApplicationHostEnvironment.DevelopmentEnvironmentName;
         }
 
@@ -77,10 +64,22 @@ namespace Milochau.Core.Infrastructure.Hosting
         /// <remarks>This method only gets host name from environment variables. It shall only be used in application initialization.</remarks>
         public static string GetCurrentHostFromEnvironmentVariables()
         {
-            return Environment.GetEnvironmentVariable($"{GenericHostConfigurationPrefix}{hostNameKey}")
-                ?? Environment.GetEnvironmentVariable($"{WebHostConfigurationPrefix}{hostNameKey}")
-                ?? Environment.GetEnvironmentVariable(hostNameKey)
+            return GetValueFromEnvironment(hostNameKey)
                 ?? ApplicationHostEnvironment.LocalHostName;
+        }
+
+        private static string GetValueFromConfiguration(IConfiguration configuration, string suffix)
+        {
+            return configuration[$"{GenericHostConfigurationPrefix}{suffix}"]
+                ?? configuration[$"{WebHostConfigurationPrefix}{suffix}"]
+                ?? configuration[suffix];
+        }
+
+        private static string GetValueFromEnvironment(string suffix)
+        {
+            return Environment.GetEnvironmentVariable($"{GenericHostConfigurationPrefix}{suffix}")
+                ?? Environment.GetEnvironmentVariable($"{WebHostConfigurationPrefix}{suffix}")
+                ?? Environment.GetEnvironmentVariable(suffix);
         }
     }
 }