+++ src/Reference Projects/Milochau.Core.Console.ReferenceProject/Models/FeatureFlags.cs
@@ -1,0 +1,7 @@
+﻿namespace Milochau.Core.Console.ReferenceProject.Models
+{
+    public enum FeatureFlags
+    {
+        ReferenceProjectTest
+    }
+}